cropaway/cropaway/Resources/python/requirements.txt
# SAM3 Server Requirements
# Install with: pip install -r requirements.txt

# PyTorch for Apple Silicon (MPS)
torch>=2.0.0
torchvision>=0.15.0

# HuggingFace Transformers (SAM3 support)
# For Apple Silicon without CUDA/Triton:
# pip install git+https://github.com/huggingface/transformers
transformers>=4.40.0

# HTTP server
flask>=3.0.0
flask-cors>=4.0.0

# Image processing
pillow>=10.0.0
numpy>=1.24.0

# COCO tools for mask operations
pycocotools>=2.0.7

# Progress bars
tqdm>=4.66.0
